requirements..txt
fpdf          1.7.2
Pillow        8.4.0